chore(arch): refactoring popular movies
