+++ src/data/PopularMovies/interfaces/index.ts
@@ -1,3 +1,1 @@
 export { PopularMovies } from './IPopularMovies';
-export { MoviesResponse } from './IMoviesResponse';
-export { ImagesConfig } from './IImagesConfig';
