+++ src/data/helpers/MoviesParser/interfaces/IImagesConfig.ts
