+++ src/data/helpers/MoviesParser/interfaces/IMovies.ts
@@ -1,0 +1,8 @@
+export interface Movie {
+  id: number;
+  posterPath: string;
+  title: string;
+  releaseDate: string;
+  overview: string;
+  backdropPath: string;
+}
