+++ src/data/helpers/MoviesParser/interfaces/IMoviesResponse.ts
