+++ src/data/helpers/MoviesParser/interfaces/index.ts
@@ -1,0 +1,3 @@
+export { Movie } from './IMovies';
+export { MoviesResponse } from './IMoviesResponse';
+export { ImagesConfig } from './IImagesConfig';
